realtime-api/pyproject.toml
[tool.poetry]
name = "people-detection-api-management"
version = "0.2.0"
description = "API to manage object detection data"
authors = ["Natan Nascimento <natanascimentom@outlook.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.scripts]
start = "app:start"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
pydantic = "^1.9.1"
python-dotenv = "^0.20.0"
pymongo = {extras = ["srv"], version = "^4.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
